Move transition logic to create-or-update view
